Moved Spring code into its own package.
