+++ src/main/java/net/arin/rdap_bootstrap/spring/AppProperties.java
@@ -13,7 +13,7 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
  * IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-package net.arin.rdap_bootstrap;
+package net.arin.rdap_bootstrap.spring;
 
 import org.springframework.core.env.PropertyResolver;
 import org.springframework.core.env.StandardEnvironment;
